--- a/examples/usage/main.go
+++ b/examples/usage/main.go
@@ -2,8 +2,9 @@ package main
 
 import (
 	"fmt"
-	"github.com/senpathi/gofloat"
 	"math"
+
+	"github.com/senpathi/gofloat"
 )
 
 func main() {