After testing the performance of the benchmarks with the go tool pprof, there was a huge impact from the math package, and mostly `math.Pow()`

So I have changed `math.Pow()` to `math.Pow10` which is a better performer for calculating 10th powers and also stores the calculated power value to avoid later calculations.

Here are the benchmark results before and after
### Before
```
BenchmarkFloat/Bench_ToFloat-10                 92825976                12.73 ns/op
BenchmarkFloat/Bench_Float_Add-10               19134381                62.89 ns/op
BenchmarkFloat/Bench_Float_Sub-10               19082767                62.88 ns/op
BenchmarkFloat/Bench_Float_Multiply-10          19102563                62.86 ns/op
BenchmarkFloat/Bench_Float_Divide-10            15634525                76.69 ns/op
```
### After
```
BenchmarkFloat/Bench_ToFloat-10                 489519688                2.285 ns/op
BenchmarkFloat/Bench_Float_Add-10               134394903                8.932 ns/op
BenchmarkFloat/Bench_Float_Sub-10               134450560                8.930 ns/op
BenchmarkFloat/Bench_Float_Multiply-10          133463366                9.000 ns/op
BenchmarkFloat/Bench_Float_Divide-10            126986404                9.446 ns/op
```